chore(ai): rename tool methods to ai
