+++ lib/services/data/shifts/ai_tool_methods/models/shift_assignments_result_model.dart
