+++ lib/services/data/shifts/ai_tool_methods/models/shift_clock_in_out_result_model.dart
