+++ lib/services/data/shifts/ai_tool_methods/models/shift_log_results_model.dart
