+++ lib/services/data/shifts/ai_tool_methods/models/shift_request_models.dart
