+++ lib/services/data/tasks/ai_tool_methods/ai_tool_execution_utils.dart
