+++ lib/services/data/tasks/ai_tool_methods/task_tool_methods.dart
@@ -24,7 +24,7 @@ import 'package:seren_ai_flutter/services/data/tasks/filtered/task_filter_view_t
 import 'package:seren_ai_flutter/services/data/tasks/filtered/tasks_filtered_search_provider.dart';
 import 'package:seren_ai_flutter/services/data/tasks/providers/task_navigation_service.dart';
 import 'package:seren_ai_flutter/services/data/tasks/task_field_enum.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/ai_date_parser.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/ai_date_parser.dart';
 import 'package:seren_ai_flutter/services/data/common/status_enum.dart';
 import 'package:seren_ai_flutter/services/data/common/utils/string_similarity_extension.dart';
 import 'package:seren_ai_flutter/services/data/common/widgets/delete_confirmation_dialog.dart';
@@ -39,17 +39,17 @@ import 'package:seren_ai_flutter/services/data/tasks/providers/cur_selected_task
 import 'package:seren_ai_flutter/services/data/tasks/providers/task_assignments_service_provider.dart';
 import 'package:seren_ai_flutter/services/data/tasks/repositories/task_comments_repository.dart';
 import 'package:seren_ai_flutter/services/data/tasks/repositories/tasks_repository.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/models/add_comment_to_task_result_model.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/models/create_task_result_model.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/models/delete_task_result_model.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/models/find_tasks_result_model.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/models/show_tasks_result_model.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/models/task_request_models.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/models/update_task_fields_result_model.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/models/add_comment_to_task_result_model.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/models/create_task_result_model.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/models/delete_task_result_model.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/models/find_tasks_result_model.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/models/show_tasks_result_model.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/models/task_request_models.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/models/update_task_fields_result_model.dart';
 import 'package:seren_ai_flutter/services/data/tasks/widgets/action_buttons/edit_task_button.dart';
 import 'package:seren_ai_flutter/services/data/users/models/user_model.dart';
 import 'package:seren_ai_flutter/services/data/users/repositories/users_repository.dart';
-import 'package:seren_ai_flutter/services/data/tasks/tool_methods/ai_tool_execution_utils.dart';
+import 'package:seren_ai_flutter/services/data/tasks/ai_tool_methods/ai_tool_execution_utils.dart';
 import 'package:seren_ai_flutter/widgets/search/search_modal.dart';
 import 'package:seren_ai_flutter/services/ai/ai_context_service.dart';
 import 'package:intl/intl.dart';
